fix: kline entity 用 rust_decimal::Decimal 匹配 MySQL DECIMAL 类型
MySQL 列是 DECIMAL(12,4)，entity 用 f64（对应 SQL DOUBLE）导致 sqlx
解码失败。改用 rust_decimal::Decimal 保持定点精度（金融数据最佳实践）。
- sea-orm 启用 with-rust_decimal feature
- kline entity 字段 f64 -> Decimal
- kline_service upsert 时 f64 -> Decimal 转换
- rust_decimal 启用 serde-float，API 仍返回 JSON 数字，前端无需改

+++ crates/datasphere-entity/src/kline.rs
@@ -1,3 +1,4 @@
+use rust_decimal::Decimal;
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, serde::Serialize)]
@@ -7,14 +8,14 @@ pub struct Model {
     pub id: i64,
     pub code: String,
     pub date: chrono::NaiveDate,
-    pub open: f64,
-    pub close: f64,
-    pub high: f64,
-    pub low: f64,
+    pub open: Decimal,
+    pub close: Decimal,
+    pub high: Decimal,
+    pub low: Decimal,
     pub volume: i64,
-    pub amount: f64,
-    pub turnover: Option<f64>,
-    pub pct_change: Option<f64>,
+    pub amount: Decimal,
+    pub turnover: Option<Decimal>,
+    pub pct_change: Option<Decimal>,
     pub created_at: chrono::NaiveDateTime,
     pub updated_at: chrono::NaiveDateTime,
 }
